refactor(provider/websocket): split server.go into server / peer / persistence (#21)
provider/websocket/server.go had grown to 956 lines mixing five concerns:
HTTP upgrade, peer lifecycle, sync-protocol dispatch, awareness
broadcast, and persistence worker logic. Split into three focused files
in the same package:

  server.go        — Server type, HTTP upgrade, room registry, Shutdown
  peer.go          — peer type, read/write loops, broadcast, handleDisconnect
  persistence.go   — startPersistenceWorker extracted from getOrCreateRoom

Pure code movement: zero behavior change, all existing tests pass
without modification, no public API change. The package remains
internally consistent — peer's references to Server fields and
server's references to peer methods cross file boundaries within
the same package, which is idiomatic Go.

inject.go and doc.go are unchanged.
