+++ provider/websocket/server.go
@@ -1,19 +1,9 @@
-// Package websocket provides a net/http-compatible WebSocket handler that
-// synchronises Yjs documents between multiple peers using the y-protocols
-// sync and awareness protocols.
-//
-// Usage:
-//
-//	srv := websocket.NewServer()
-//	http.Handle("/yjs/{room}", srv)
-//	http.ListenAndServe(":8080", nil)
 package websocket
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"log"
 	"log/slog"
 	"net/http"
 	"net/url"
@@ -151,28 +141,6 @@ type room struct {
 	persistCh   chan []byte   // buffered channel for serialised writes
 	persistStop chan struct{} // closed to signal goroutine to drain and exit
 	persistDone chan struct{} // closed when persistence goroutine exits
-}
-
-// peer is one connected WebSocket client.
-type peer struct {
-	conn       *gws.Conn
-	wmu        sync.Mutex // serialises concurrent writes
-	closed     bool       // H2: true after handleDisconnect; guarded by wmu
-	room       *room
-	roomName   string              // C1: name used to delete room when empty
-	server     *Server             // C1: back-reference for room map cleanup
-	done       chan struct{}       // H1: closed when the read loop exits
-	clientIDs  map[uint64]struct{} // awareness clientIDs controlled by this peer
-	cidMu      sync.Mutex
-	writeCh    chan []byte   // buffered queue drained by runWriter goroutine
-	writerDone chan struct{} // closed when runWriter exits
-
-	// disconnectOnce ensures the full teardown sequence in handleDisconnect
-	// runs exactly once, regardless of how many callers race (e.g. broadcast's
-	// conn.Close() triggering the read loop while a ctx-cancel path also calls
-	// handleDisconnect). The closed bool under wmu is still needed for the
-	// per-operation guards in broadcast/write/runWriter.
-	disconnectOnce sync.Once
 }
 
 // Server is a net/http-compatible WebSocket handler.
@@ -429,45 +397,7 @@ func (s *Server) getOrCreateRoom(name string) (*room, error) {
 		r.persistCh = make(chan []byte, 256)
 		r.persistStop = make(chan struct{})
 		r.persistDone = make(chan struct{})
-		go func() {
-			defer close(r.persistDone)
-			store := func(update []byte) {
-				defer func() {
-					if rv := recover(); rv != nil {
-						log.Printf("ygo/websocket: StoreUpdate panic for room %q: %v", name, rv)
-					}
-				}()
-				if err := s.persistence.StoreUpdate(name, update); err != nil {
-					log.Printf("ygo/websocket: StoreUpdate for room %q: %v", name, err)
-				}
-			}
-			for {
-				select {
-				case update := <-r.persistCh:
-					store(update)
-				case <-r.persistStop:
-					// Drain buffered updates before exiting.
-					for {
-						select {
-						case update := <-r.persistCh:
-							store(update)
-						default:
-							return
-						}
-					}
-				case <-s.shutdownCh:
-					// Server is shutting down; drain any remaining buffered updates.
-					for {
-						select {
-						case update := <-r.persistCh:
-							store(update)
-						default:
-							return
-						}
-					}
-				}
-			}
-		}()
+		s.startPersistenceWorker(r, name)
 		r.doc.OnUpdate(func(update []byte, _ any) {
 			select {
 			case r.persistCh <- update:
@@ -627,330 +557,4 @@ func encodeSyncStep2Msg(update []byte) []byte {
 	enc.WriteVarUint(ygsync.MsgSyncStep2)
 	enc.WriteVarBytes(update)
 	return enc.Bytes()
-}
-
-// handleMessage decodes the outer message type and dispatches accordingly.
-func (p *peer) handleMessage(data []byte) {
-	dec := encoding.NewDecoder(data)
-	outerType, err := dec.ReadVarUint()
-	if err != nil {
-		return
-	}
-
-	switch outerType {
-	case msgSync:
-		// Sync payload follows directly (no VarBytes wrapper).
-		payload := dec.RemainingBytes()
-		reply, err := ygsync.ApplySyncMessage(p.room.doc, payload, p)
-		if err != nil {
-			return
-		}
-		if reply != nil {
-			// Peer sent step-1 — send step-2 reply only to them.
-			p.sendSync(reply)
-		} else {
-			// Peer sent step-2 or update — broadcast to all other peers.
-			p.broadcastSync(payload)
-		}
-
-	case msgAwareness:
-		// Awareness payload is VarBytes-wrapped (y-websocket protocol).
-		awBytes, err := dec.ReadVarBytes()
-		if err != nil {
-			return
-		}
-		p.trackAwarenessClients(awBytes)
-		if err := p.room.awareness.ApplyUpdate(awBytes, p); err != nil {
-			return // Drop invalid awareness updates; do not broadcast.
-		}
-		p.broadcastAwareness(awBytes)
-
-	case msgAuth:
-		// Auth messages (type 2) are defined by y-websocket but not used by
-		// this server. Silently ignore.
-
-	case msgQueryAwareness:
-		p.sendAwareness(p.room.awareness.EncodeUpdate(nil))
-	}
-}
-
-// trackAwarenessClients records which awareness clientIDs this peer owns
-// so they can be removed when the peer disconnects.
-func (p *peer) trackAwarenessClients(payload []byte) {
-	dec := encoding.NewDecoder(payload)
-	n, err := dec.ReadVarUint()
-	if err != nil {
-		return
-	}
-	p.cidMu.Lock()
-	defer p.cidMu.Unlock()
-	for i := uint64(0); i < n; i++ {
-		clientID, err := dec.ReadVarUint()
-		if err != nil {
-			return
-		}
-		if _, err = dec.ReadVarUint(); err != nil { // clock
-			return
-		}
-		if _, err = dec.ReadVarString(); err != nil { // state JSON
-			return
-		}
-		// Cap the number of clientIDs a single peer may claim to prevent OOM
-		// when handleDisconnect builds the removal slice (N-H4).
-		if len(p.clientIDs) < maxAwarenessClientsPerPeer {
-			p.clientIDs[clientID] = struct{}{}
-		}
-	}
-}
-
-// handleDisconnect removes the peer from the room and broadcasts awareness
-// removal for all clientIDs the peer owned.
-//
-// disconnectOnce ensures the full teardown body runs at most once even when
-// multiple callers race — e.g. broadcast()'s conn.Close() waking the read loop
-// while the ctx-cancel goroutine concurrently reaches this point.
-func (p *peer) handleDisconnect() {
-	p.disconnectOnce.Do(func() {
-		// H2: mark closed so concurrent broadcast writes skip this peer.
-		// Close writeCh after marking closed so runWriter can drain and exit.
-		// Both operations are done under wmu so broadcast() sees a consistent
-		// state (closed=true is visible before writeCh is closed).
-		p.wmu.Lock()
-		p.closed = true
-		close(p.writeCh)
-		p.wmu.Unlock()
-
-		// Wait for the per-peer writer goroutine to fully exit before we touch
-		// the connection in the teardown path. The writer will see the closed
-		// channel and exit cleanly.
-		<-p.writerDone
-
-		rm := p.room
-
-		// Acquire both locks (server map first, then room) to atomically remove
-		// the peer and, if the room is now empty, delete the room from the server
-		// map and stop the persistence goroutine. This prevents a TOCTOU race
-		// where a new peer joins between the emptiness check and room deletion,
-		// which would fork the logical document into two rooms.
-		p.server.rmu.Lock()
-		rm.mu.Lock()
-		delete(rm.peers, p)
-		empty := len(rm.peers) == 0
-		if empty {
-			delete(p.server.rooms, p.roomName)
-			if rm.persistStop != nil {
-				close(rm.persistStop)
-			}
-		}
-		rm.mu.Unlock()
-		p.server.rmu.Unlock()
-
-		// Release semaphore slots now that the peer has left.
-		if rm.peerSem != nil {
-			rm.peerSem.Release(1)
-		}
-		if sem := p.server.connSemaphore(); sem != nil {
-			sem.Release(1)
-		}
-
-		// Wait for the persistence goroutine to drain buffered writes before the
-		// room reference becomes garbage. This runs outside the locks above.
-		if empty && rm.persistDone != nil {
-			<-rm.persistDone
-		}
-
-		p.cidMu.Lock()
-		clientIDs := make([]uint64, 0, len(p.clientIDs))
-		for id := range p.clientIDs {
-			clientIDs = append(clientIDs, id)
-		}
-		p.cidMu.Unlock()
-
-		if len(clientIDs) == 0 {
-			return
-		}
-
-		removalBytes := encodeAwarenessRemoval(p.room.awareness, clientIDs)
-		if removalBytes == nil {
-			return
-		}
-		if err := p.room.awareness.ApplyUpdate(removalBytes, nil); err != nil {
-			p.server.log().Warn("apply removal awareness failed", "room", p.roomName, "err", err)
-		}
-		p.broadcastAwarenessFromRoom(removalBytes)
-	})
-}
-
-// encodeAwarenessRemoval builds a raw awareness update that marks the given
-// client IDs as removed (null state, clock incremented by 1).
-func encodeAwarenessRemoval(aw *awareness.Awareness, clientIDs []uint64) []byte {
-	states := aw.GetStates()
-	var toRemove []struct {
-		id    uint64
-		clock uint64
-	}
-	for _, id := range clientIDs {
-		if cs, ok := states[id]; ok {
-			toRemove = append(toRemove, struct {
-				id    uint64
-				clock uint64
-			}{id, cs.Clock})
-		}
-	}
-	if len(toRemove) == 0 {
-		return nil
-	}
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(uint64(len(toRemove)))
-	for _, item := range toRemove {
-		enc.WriteVarUint(item.id)
-		enc.WriteVarUint(item.clock + 1)
-		enc.WriteVarString("null")
-	}
-	return enc.Bytes()
-}
-
-// sendSync writes a sync message (outer type 0, raw payload) to this peer.
-func (p *peer) sendSync(syncMsg []byte) {
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(msgSync)
-	enc.WriteRaw(syncMsg) // sync payload is NOT VarBytes-wrapped
-	p.write(enc.Bytes())
-}
-
-// sendAwareness writes an awareness message (outer type 1, VarBytes payload)
-// to this peer.
-func (p *peer) sendAwareness(awMsg []byte) {
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(msgAwareness)
-	enc.WriteVarBytes(awMsg) // awareness payload IS VarBytes-wrapped
-	p.write(enc.Bytes())
-}
-
-// broadcastSync sends a sync message to all OTHER peers in the room.
-func (p *peer) broadcastSync(syncMsg []byte) {
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(msgSync)
-	enc.WriteRaw(syncMsg)
-	p.broadcast(enc.Bytes(), true)
-}
-
-// broadcastAwareness sends an awareness message to all OTHER peers in the room.
-func (p *peer) broadcastAwareness(awMsg []byte) {
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(msgAwareness)
-	enc.WriteVarBytes(awMsg)
-	p.broadcast(enc.Bytes(), true)
-}
-
-// broadcastAwarenessFromRoom sends an awareness message to ALL peers (called
-// from disconnect handler which has already removed itself from the room).
-func (p *peer) broadcastAwarenessFromRoom(awMsg []byte) {
-	enc := encoding.NewEncoder()
-	enc.WriteVarUint(msgAwareness)
-	enc.WriteVarBytes(awMsg)
-	p.broadcast(enc.Bytes(), false)
-}
-
-// broadcast enqueues data for delivery to peers in the room. If excludeSelf
-// is true, the calling peer is excluded.
-//
-// When a target peer's writeCh is full (slow peer, dead connection, or
-// receiver lagging), the peer is disconnected rather than dropping the
-// message. This matches Rust yrs-warp's bounded-broadcast pattern: a
-// dropped message would leave the peer with a silent gap in their sync
-// stream that only resolves on the next exchange. Disconnecting forces
-// a reconnect-and-resync flow which the CRDT's pending-structs
-// machinery handles cleanly.
-func (p *peer) broadcast(data []byte, excludeSelf bool) {
-	p.room.mu.Lock()
-	targets := make([]*peer, 0, len(p.room.peers))
-	for other := range p.room.peers {
-		if excludeSelf && other == p {
-			continue
-		}
-		targets = append(targets, other)
-	}
-	p.room.mu.Unlock()
-
-	for _, other := range targets {
-		// Guard against sending to a closed channel: check p.closed under
-		// wmu before attempting the channel send. handleDisconnect sets closed
-		// under wmu before closing writeCh, so this is race-free.
-		other.wmu.Lock()
-		if other.closed {
-			other.wmu.Unlock()
-			continue
-		}
-		select {
-		case other.writeCh <- data:
-			// queued
-		default:
-			// Queue full — disconnect the slow peer.
-			p.server.log().Warn("peer write queue full; closing slow peer",
-				"room", other.roomName,
-				"queueSize", cap(other.writeCh))
-			_ = other.conn.Close()
-			// The peer's read loop will detect the close and run
-			// handleDisconnect to clean up room state.
-		}
-		other.wmu.Unlock()
-	}
-}
-
-// write enqueues a raw binary message for delivery to this peer via the
-// per-peer writer goroutine. H2: skips the write if the peer has already
-// been marked closed. If the queue is full the peer is disconnected:
-// a dropped handshake reply (sendSync / sendAwareness) leaves the remote
-// peer hung waiting for a response it will never receive. Closing the
-// connection forces a reconnect-and-resync, matching the broadcast contract.
-func (p *peer) write(data []byte) {
-	p.wmu.Lock()
-	if p.closed {
-		p.wmu.Unlock()
-		return
-	}
-	select {
-	case p.writeCh <- data:
-	default:
-		// Queue full during a direct send (e.g. handshake reply) — disconnect.
-		// Unlike a silent drop, closing the connection lets the CRDT
-		// pending-structs machinery recover via reconnect-and-resync.
-		p.server.log().Warn("peer write queue full during direct send; closing peer",
-			"room", p.roomName,
-			"queueSize", cap(p.writeCh))
-		_ = p.conn.Close()
-	}
-	p.wmu.Unlock()
-}
-
-// runWriter is the dedicated per-peer broadcast writer. It drains writeCh
-// and serialises writes to the connection. Exits when writeCh is closed
-// (during teardown) or when a write fails (the connection is then dead;
-// the read loop will tear down the peer).
-//
-// This pattern mirrors Rust yrs-warp's per-peer sink task. It replaces
-// the previous "spawn one goroutine per peer per broadcast" model, which
-// produced unbounded goroutine churn under high broadcast cardinality
-// and had no backpressure mechanism.
-func (p *peer) runWriter() {
-	defer close(p.writerDone)
-	for data := range p.writeCh {
-		p.wmu.Lock()
-		if p.closed {
-			p.wmu.Unlock()
-			return
-		}
-		if err := p.conn.SetWriteDeadline(time.Now().Add(writeTimeout)); err != nil {
-			p.server.log().Debug("set write deadline failed", "err", err)
-			p.wmu.Unlock()
-			return
-		}
-		if err := p.conn.WriteMessage(gws.BinaryMessage, data); err != nil {
-			p.server.log().Warn("write to peer failed; closing", "room", p.roomName, "err", err)
-			p.wmu.Unlock()
-			return
-		}
-		p.wmu.Unlock()
-	}
 }
